feat; create screen to transfer LSP7 tokens

+++ packages/reactnative/src/hooks/useTokenBalance.ts
@@ -34,7 +34,7 @@ export default function useTokenBalance({
   }, [tokenAddress, account.address, refetch]);
 
   return {
-    balance: data ? data : null, // Assuming balanceOf returns a tuple [balance]
+    balance: data ? (data as bigint) : null, // Assuming balanceOf returns a tuple [balance]
     isLoading,
     error
   };
